+++ ops-center/frontend/src/App.vue
@@ -20,6 +20,10 @@
           <el-icon><Switch /></el-icon>
           <span>功能开关</span>
         </el-menu-item>
+        <el-menu-item index="/runtime-flags">
+          <el-icon><SwitchButton /></el-icon>
+          <span>桌面端功能开关</span>
+        </el-menu-item>
         <el-menu-item index="/secrets">
           <el-icon><Key /></el-icon>
           <span>Key 管理</span>